Fix bad circular reference in Node.js streams
This has apparently been broken for a while but no one noticed.
